refactor(providers): extract OneConsole cookie-header pair from Alibaba + Qwen copies
AlibabaTokenPlanCookieHeaders and QwenCloudCookieHeaders are byte-for-byte identical apart from the two cached-header name constants. The HTTPCookie -> scoped-Cookie-header builder (matchesRequestURL, cookieSortKey, by-name merge) and the duplicated '[String].uniquedSorted()' extension also live in both provider files.

Promote a generic OneConsoleCookieHeaders and OneConsoleCookieHeaderBuilder under Providers/Shared/AliyunOneConsole/. Each provider declares its cache namespace (alibaba_token_plan, qwen_cloud) via a typed initializer on the shared struct and routes cached-header round-trips through it.

AlibabaTokenPlanCookieHeaders and QwenCloudCookieHeaders become typealiases of OneConsoleCookieHeaders, so all existing call sites continue to compile. The cached-header initializer argument is now namespaced (alibabaTokenPlanCachedHeader:, qwenCloudCachedHeader:) to avoid the collision that would otherwise require the same init? across two extensions.

Drop the three duplicate '[String].uniquedSorted()' fileprivate extensions (AlibabaTokenPlanProviderDescriptor, AlibabaTokenPlanUsageFetcher, QwenCloudProviderDescriptor, QwenCloudUsageFetcher) now that the shared type owns one copy.

No behavior change.

Verified:
- swift test --filter AlibabaCodingPlan: 58 tests passed
- swift test --filter AlibabaTokenPlan: 37 tests passed
- swift test --filter QwenCloud: 27 tests passed
- make check: 0 violations
