+++ Tests/CodexBarTests/QwenCloudProviderTests.swift
@@ -314,7 +314,7 @@ struct QwenCloudCookieHeaderTests {
             apiCookieHeader: "login_aliyunid_ticket=ticket; api_only=api",
             dashboardCookieHeader: "login_aliyunid_ticket=ticket; dashboard_only=dashboard")
 
-        let cached = try #require(QwenCloudCookieHeaders(cachedHeader: headers.cacheCookieHeader))
+        let cached = try #require(QwenCloudCookieHeaders(qwenCloudCachedHeader: headers.cacheQwenCloudCookieHeader()))
 
         #expect(cached.apiCookieHeader.contains("api_only=api"))
         #expect(!cached.apiCookieHeader.contains("dashboard_only=dashboard"))
